fix(Page session): put save exercise button at the end for better ux
