+++ app/session/page.jsx
@@ -5,18 +5,18 @@ import { useRouter } from "next/navigation";
 import classes from "./session.module.css";
 import SessionContext from "@modules/client/contexts/sessionProvider";
 import { Accordion, AccordionItem, Avatar, Image } from "@heroui/react";
-import useStopwatch from "@modules/client/utils/useStopwatch";
 import useTimer from "@modules/client/utils/useTimer";
 import InputField from "@core/ui/Fields/InputField/InputField";
 import SelectField from "@core/ui/Fields/SelectField/SelectField";
 import BasicButton from "@core/ui/Button/BasicButton";
 import PopupButton from "@core/ui/Button/PopupButton";
 import Icon from "@core/ui/Icons/Icon";
-import { useSession } from "next-auth/react";
-import useUser from "@modules/client/requests/useUser";
 import useExercises from "@modules/client/requests/useExercises";
 import Skeleton from "@core/ui/Skeleton/Skeleton";
 import ClipLoader from "react-spinners/ClipLoader";
+import StopwatchButton from "@components/CtaButton/StopwatchButton";
+import ResetButton from "@components/CtaButton/ResetButton";
+import SaveButton from "@components/CtaButton/SaveButton";
 
 function Session() {
   const {
@@ -25,45 +25,19 @@ function Session() {
     handleOnChangeSession,
     handleAddSets,
     handleOnchangeSets,
-    refreshExercise,
+    resetExercise,
     exercisesId,
     setExercisesId,
   } = useContext(SessionContext);
   const { exercises, isLoading } = useExercises(exercisesId, "exercise");
   const [accordionKey, setAccordionKey] = useState(new Set(["1"]));
   const [isPending, startTransition] = useTransition();
-  const { time, getSeconds, getMinutes, isRunning, start, pause, reset } =
-    useStopwatch(false, session);
   const { startTimer, getFormattedTime, timers, resetTimers } =
     useTimer(session);
-  const { data: userSession, status } = useSession();
-  const { userId } = useUser(userSession);
 
   const baseUrl = process.env.NEXT_PUBLIC_API_URL;
   const cloudinaryUrl = `https://res.cloudinary.com/${process.env.NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME}`;
   const router = useRouter();
-
-  const saveExercise = async (i) => {
-    try {
-      const url = `${baseUrl}/api/stats`;
-      delete session[i].isFinished;
-      const response = await fetch(
-        url,
-        {
-          method: "POST",
-          body: JSON.stringify({
-            ...session[i],
-            rest_time: session[i].restTime,
-            training_time: session[i].trainingTime,
-            user: userId,
-          }),
-        },
-        { next: { revalidate: 10 } }
-      );
-    } catch (error) {
-      throw error;
-    }
-  };
 
   useEffect(() => {
     const session = localStorage.getItem("session");
@@ -160,100 +134,17 @@ function Session() {
             >
               <div className={classes.session_container}>
                 <div className={classes.stopwatch_buttons}>
-                  {/* Start stopwatch */}
-                  <BasicButton
-                    onAction={() => {
-                      if (isRunning[i]) {
-                        pause(i);
-                      } else {
-                        start(i);
-                      }
-                    }}
-                    buttonContent={
-                      <div>
-                        {time[0] === 0 && time[1] === 0 ? (
-                          "Start"
-                        ) : (
-                          <>
-                            <span>
-                              {getMinutes(i).toString().padStart(2, "0")}
-                            </span>
-                            :
-                            <span>
-                              {getSeconds(i).toString().padStart(2, "0")}
-                            </span>
-                          </>
-                        )}
-                      </div>
-                    }
-                    startContent={
-                      isRunning[i] ? (
-                        <Icon
-                          name="Pause"
-                          size={16}
-                          color="white"
-                          strokeWidth={3}
-                        />
-                      ) : (
-                        <Icon
-                          name="Play"
-                          size={16}
-                          color="white"
-                          strokeWidth={3}
-                        />
-                      )
-                    }
-                    buttonStyle={`${classes.button} ${classes.start_button}`}
-                    isDisabled={exercise.isFinished}
+                  <StopwatchButton
+                    session={session}
+                    i={i}
+                    exercise={exercise}
                   />
-                  {/* Save exercise */}
-                  <PopupButton
-                    triggerAction={() => {
-                      pause(i);
-                    }}
-                    triggerButtonContent="Save"
-                    startContent={
-                      <Icon
-                        name="Save"
-                        size={16}
-                        color="white"
-                        strokeWidth={2}
-                      />
-                    }
-                    onCancel={() => start(i)}
-                    onConfirm={() => {
-                      handleOnChangeSession("trainingTime", time[i], i);
-                      saveExercise(i);
-                      handleOnChangeSession("isFinished", true, i);
-                    }}
-                    buttonStyle={`${classes.button} ${classes.save_button}`}
-                    title="Are you sure you want to end this exercise?"
-                    content="Once confirmed, this exercise will be marked as complete permanently, with no option to change the data."
-                    isDisabled={exercise.isFinished}
-                    confirmButton="Save"
-                    confirmButtonStyle={classes.save_button}
-                  />
-                  {/* refresh exercise */}
-                  <PopupButton
-                    triggerAction={() => {
-                      pause(i);
-                      handleOnChangeSession("trainingTime", time[i], i);
-                    }}
-                    triggerButtonContent={
-                      <Icon name="RefreshCcw" size={16} color="white" />
-                    }
-                    onCancel={() => start(i)}
-                    onConfirm={() => {
-                      reset(i);
-                      refreshExercise(i);
-                    }}
-                    buttonStyle={`${classes.button} ${classes.reset_button}`}
-                    isIconOnly={true}
-                    title="Are you sure you want to reset this exercise?"
-                    content="Confirming will erase all progress and restart this exercise from zero."
-                    isDisabled={exercise.isFinished}
-                    confirmButton="Reset"
-                    confirmButtonStyle={classes.reset_button}
+                  <ResetButton
+                    session={session}
+                    handleOnChangeSession={handleOnChangeSession}
+                    i={i}
+                    exercise={exercise}
+                    resetExercise={resetExercise}
                   />
                 </div>
                 {/* Choose RM */}
@@ -777,7 +668,13 @@ function Session() {
                     );
                   })}
                 </div>
-                {/* stopwatch buttons */}
+                {/* save button */}
+                <SaveButton
+                  handleOnChangeSession={handleOnChangeSession}
+                  session={session}
+                  i={i}
+                  exercise={exercise}
+                />
               </div>
             </AccordionItem>
           );
