+++ components/CtaButton/ctaButton.module.css
@@ -1,0 +1,20 @@
+.button {
+  display: flex;
+  flex-direction: row;
+  width: 100%;
+  background-color: var(--primary-accent-color);
+  color: var(--primary-light-color);
+  font-size: var(--font-xl);
+  font-weight: var(--font-black);
+  font-style: italic;
+  text-transform: uppercase;
+  padding: 0;
+}
+
+.save_button {
+  background-color: var(--success-color);
+}
+
+.reset_button {
+  background-color: var(--danger-color);
+}
